@@ -14,22 +14,23 @@
   "dependencies": {
     "@phosphor-icons/react": "^2.1.5",
     "decimal.js": "^10.4.3",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
+    "react-intl": "^6.6.6",
     "react-modal": "^3.16.1",
+    "react-router": "^6.23.1",
+    "react-router-dom": "^6.23.1",
     "react-tooltip": "^5.26.3",
     "react-use": "^17.5.0"
   },
   "devDependencies": {
-    "@types/react": "^17.0.4",
-    "@types/react-dom": "^17.0.3",
+    "@types/react": "^18.3.2",
+    "@types/react-dom": "^18.3.0",
     "@types/react-modal": "^3.16.3",
-    "@types/react-tooltip": "^4.2.4",
     "@types/webpack": "^5.28.5",
-    "@types/webpack-dev-server": "^4.7.2",
     "@types/webpack-env": "^1.18.4",
     "copy-webpack-plugin": "^12.0.2",
-    "css-loader": "^6.10.0",
+    "css-loader": "^7.1.1",
     "file-loader": "^6.2.0",
     "gh-pages": "^6.1.1",
     "html-webpack-plugin": "^5.6.0",
@@ -39,7 +40,7 @@
     "stylus-loader": "^8.1.0",
     "ts-loader": "^9.5.1",
     "ts-node": "^10.9.2",
-    "typescript": "^4.2.4",
+    "typescript": "^5.4.5",
     "webpack": "^5.90.3",
     "webpack-cli": "^5.1.4",
     "webpack-dev-server": "^5.0.3"
