@@ -1,34 +1,29 @@
 import React, { useState, useEffect } from 'react';
 import Modal from "react-modal"
-import Simple from "components/pages/simple";
-import { Tooltip } from "react-tooltip";
-import ItemTooltip from 'components/tooltip/item-tooltip';
-import SubjectSkillTooltip from 'components/tooltip/subject-skill/subject-skill-tooltip';
+import Simple from "./pages/simple";
 import Navigation from 'components/pages/navigation/navigation';
-import style from "./App.styl";
+import { Route, Routes } from 'react-router';
+import { IntlProvider } from 'react-intl';
+import { locales } from './intl';
 
 interface AppProps {}
 
 
 function App({}: AppProps) {
-  React.useLayoutEffect(() => {
-    Modal.setAppElement("#root");
-  });
+    React.useLayoutEffect(() => {
+        Modal.setAppElement("#root");
+    });
 
-  // Create the count state.
-  const [count, setCount] = useState(0);
-  // Create the counter (+1 every second).
-  useEffect(() => {
-    const timer = setTimeout(() => setCount(count + 1), 1000);
-    return () => clearTimeout(timer);
-  }, [count, setCount]);
-  // Return the App component.
-  return (
-    <>
-      <Navigation />
-      <Simple />
-    </>
-  );
+    return (
+        <IntlProvider locale={"ja"} messages={locales["ja"]} >
+            <Navigation />
+            <Routes>
+                <Route path="/" element={<Simple />} />
+                <Route path="/simple" element={<Simple />} />
+                <Route path="/combat" element={<Simple />} />
+            </Routes>
+        </IntlProvider>
+    );
 }
 
 export default App;
