@@ -5,6 +5,7 @@ import {
 } from "@phosphor-icons/react"
 import { styles } from "@app/util/style";
 import { useWindowSize } from "react-use";
+import { Link } from "react-router-dom";
 
 type Props = {
 
@@ -18,14 +19,14 @@ const navigation: React.FC<Props> = props => {
         <nav className={styles(style.nav, width < 1400 ? style.small : undefined)}>
             <h1>ルミア計算機</h1>
             <hr />
-            <div className={style.button}>
+            <Link to="/simple" className={style.button}>
                 <Calculator />
                 <h2>シンプル</h2>
-            </div>
-            <div className={style.button}>
+            </Link>
+            <Link to="/combat" className={style.button}>
                 <Sword />
-                <h2>対戦(作成中)</h2>
-            </div>
+                <h2>対戦</h2>
+            </Link>
             <div className={style.button}>
                 <ChartBar />
                 <h2>成長(作成中)</h2>
@@ -45,10 +46,10 @@ const navigation: React.FC<Props> = props => {
             </a>
             <hr />
             <div className={style.acknowledgements}>
-                <a href="https://playeternalreturn.com" target="_blank">エターナルリターン</a>はNimble Neuron社の著作物です。<br />
+                <a className={style.link} href="https://playeternalreturn.com" target="_blank">エターナルリターン</a>はNimble Neuron社の著作物です。<br />
                 <br />
                 ルミア計算機は個人の研究のため作成された非営利・非公式Webサイトです。<br />
-                画像は<a href="https://drive.google.com/drive/folders/1bgW32L09YPpRgQKtH4C_TAd3Kr0N9Y90" target="_blank">ファンキット</a>より引用され、ゲーム内テキストおよび文章はエターナルリターンより引用されています。
+                画像は<a className={style.link} href="https://drive.google.com/drive/folders/1bgW32L09YPpRgQKtH4C_TAd3Kr0N9Y90" target="_blank">ファンキット</a>より引用され、ゲーム内テキストおよび文章はエターナルリターンより引用されています。
             </div>
         </nav>
         </IconContext.Provider>
