@@ -2,11 +2,10 @@ import * as React from "react";
 import useBaseValue from "./use-base-value";
 import BaseValue from "./base-value";
 import Additional from "./additional-value";
-import { DisplayedValues, MasteryValues } from "../use-status";
-import MultipliedValue from "./multipled-value";
 import Decimal from "decimal.js";
 import InnerTable from "components/common/inner-table";
 
+/*
 type Props = MasteryValues & {
     perLevel?: Decimal
     level: number
@@ -31,3 +30,4 @@ const basicAttackAmp: React.FC<Props> = props => {
 }
 
 export default basicAttackAmp;
+*/
