@@ -1,23 +1,36 @@
-import { StatusProps } from "components/subject/status";
 import Constants from "./constants.json";
-import Decimal from "decimal.js";
-import { SubjectConfig } from "components/subject/use-subject-config";
+import Decimal from "decimal.js";import { SubjectConfig } from "app-types/subject-dynamic/config";
+import { StatusOverrideFunc } from "../status-override";
+import { Status } from "app-types/subject-dynamic/status/type";
+;
 
-export function additionalAmp(status: StatusProps, config: SubjectConfig): Decimal {
+export function additionalAmp(status: Status, config: SubjectConfig): Decimal {
     const additionalAS = status.attackSpeed.base.times(status.attackSpeed.multiplier);
     const ratio = new Decimal(Constants.T.amp_per_as[config.skillLevels.T]);
     return additionalAS.times(ratio);
 }
 
-export default function(status: StatusProps, config: SubjectConfig): StatusProps {
-    return {
-        ...status,
-        baseSkillAmp: status.baseSkillAmp.add(additionalAmp(status, config)),
-        attackSpeed: {
-            base: new Decimal(Constants.T.attack_speed),
-            multiplier: status.attackSpeed.multiplier,
-            calculated: new Decimal(Constants.T.attack_speed)
-        },
-        basicAttackRange: status.basicAttackRange.add(Constants.T.additional_attack_range)
-    };
-}
+const f: StatusOverrideFunc = (status, config) => ({
+    ...status,
+    skillAmp: {
+        ...status.skillAmp,
+        overrideAdditional: {
+            nameKey: "subject.adela.passive-amp",
+            value: additionalAmp(status, config)
+        }
+    },
+    attackSpeed: {
+        ...status.attackSpeed,
+        base: new Decimal(Constants.T.attack_speed),
+        value: new Decimal(Constants.T.attack_speed)
+    },
+    basicAttackRange: {
+        ...status.basicAttackRange,
+        overrideAdditional: {
+            nameKey: "subject.adela.passive-range",
+            value: new Decimal(Constants.T.additional_attack_range)
+        }
+    }
+})
+
+export default f;
