@@ -1,7 +1,7 @@
 import { StatusProps } from "components/subject/status";
 import Constants from "./constants.json";
-import { SubjectConfig } from "components/subject/use-subject-config";
 import Decimal from "decimal.js";
+import { SubjectConfig } from "app-types/subject-dynamic/config";
 
 export function additionalAmp(status: StatusProps): Decimal {
     return status.cooldownReduction.times(Constants.T.cooldown_conversion);
