@@ -1,7 +1,7 @@
 import { StatusProps } from "components/subject/status";
 import Constants from "./constants.json";
 import Decimal from "decimal.js";
-import { SubjectConfig } from "components/subject/use-subject-config";
+import { SubjectConfig } from "app-types/subject-dynamic/config";
 
 export default function(status: StatusProps, config: SubjectConfig): StatusProps {
     return {
