@@ -1,5 +1,5 @@
+import { SubjectConfig } from "app-types/subject-dynamic/config";
 import { Status } from "components/subject/status";
-import { SubjectConfig } from "components/subject/use-subject-config";
 
 export type SubjectSkillProps = {
     showEquation: boolean
