@@ -1,7 +1,6 @@
 import { StatusProps } from "components/subject/status";
 import Constants from "./constants.json";
-import Decimal from "decimal.js";
-import { SubjectConfig } from "components/subject/use-subject-config";
+import { SubjectConfig } from "app-types/subject-dynamic/config";
 
 export default function(status: StatusProps, config: SubjectConfig): StatusProps {
     const additionalMaxHP = config.stack * Constants.T.max_hp[config.skillLevels.T];
