@@ -2,7 +2,6 @@ import * as path from "path";
 import * as webpack from "webpack";
 import "webpack-dev-server";
 const HtmlWebpackPlugin = require("html-webpack-plugin");
-const CopyWebpackPlugin = require("copy-webpack-plugin");
 
 const config: webpack.Configuration = {
     entry: {
@@ -27,11 +26,15 @@ const config: webpack.Configuration = {
             {
                 test: /\.styl$/,
                 use: [
-                    "style-loader",
+                    {
+                        loader: "style-loader"
+                    },
                     {
                         loader: "css-loader",
                         options: {
                             modules: {
+                                namedExport: false,
+                                exportLocalsConvention: "as-is",
                                 localIdentName: "[path][name]-[local]"
                             }
                         }
@@ -49,14 +52,16 @@ const config: webpack.Configuration = {
     ],
     devServer: {
         host: "0.0.0.0",
-        port: 8080
+        port: 8080,
+        historyApiFallback: true
     },
     resolve: {
-        extensions: [".ts", ".tsx", ".js", ".jsx", ".png"],
+        extensions: [".ts", ".tsx", ".js", ".jsx", ".png", ".styl"],
         modules: [path.resolve(__dirname, "src"), "node_modules"],
         alias: {
             "@app": path.resolve(__dirname, "./src"),
-            "dict": path.resolve(__dirname, "./src/dictionary"),
+            "dictionary": path.resolve(__dirname, "./src/dictionary-jsons"),
+            "app-types": path.resolve(__dirname, "./src/types/app-types"),
             "components": path.resolve(__dirname, "./src/components"),
             "resources": path.resolve(__dirname, "../resources"),
             "util": path.resolve(__dirname, "./src/util")
